Possible fix for hang in Windows when scrolling with arrow key.
9ce4c71449ba68baffdba421edc3be9e28d435af changed iread on Windows
so that if WIN32kbhit detects a keypress, it ungets the key and
returns EOI. The problem is, the next call to iread will then
detect that same queued key, unget it, and immediately return,
so we are stuck in a loop.

Change iread so that it looks only for a non-queued key; that is
a newly pressed key from the terminal's input.

Related to #607.