Stop trying to fill initial screen when a command is entered.
We try to fill the initial screen before accepting a command.
If a ctrl-X or other command is entered during that time, we
would previously ignore it and continue to try to fill the screen.
This commit changes the behavior so that a command will interrupt
the process of filling the screen.

Related to #608.