With -F, allow a command to interrupt get_one_screen.
When -F is in effect, get_one_screen is called at startup to
determine if the file actually fits on one screen. Allow a keyboard
command to interrupt this, but only after 3 seconds (or
LESS_SCREENFILL_TIME milliseconds) have passed. This is a compromise
between fixing #557/#559 (which wants no keyboard polling when -F
is in effect) and #608 (which wants a slow pipe with -F to be
interruptible by a keyboard command).