Improve behavior of OSC 8 ^O^N / ^O^P commands.
Previously if you used ^O^N to highlight an OSC 8 link,
then scrolled away from that page and entered ^O^N again,
the behavior was rather confusing. If there was another link
on the same line as the previously selected link, it would
return to that page and select the next link on the same line,
but if there was only one link on that line, it would search
for the first link in the currently displayed page.

This change makes the behavior more consistent and predictable.
It always search for links on the currently displayed page.
If you select a link and then scroll away from that page,
searching again will always search on the displayed page,
not the one containing the selected link. Also, if you search
for a link and then search for another link on the same page,
it will just highlight the next link, without scrolling.
It will scroll the link to the -j position only if the new
link is not on the current page.