Fix bugs introduced in 6f1a946a23d3ded8e5449c8241b203d02b6c29e2.
If we exit the loop in forw() due to reaching EOF, the final
add_forw_pos should store the position after the last line,
not the useless NULL_POSITION returned from the final forw_line.

Also, if forw_line reads a line longer than the screen width,
it should return in *p_linepos the position of the start of the
screen line, not the raw line.