Use separate bit S_SWINTERRUPT instead of S_INTERRUPT for sw-generated interrupts.
iread() sets a bit in sigs to indicate that the current operation
should be aborted. Previously it used S_INTERRUPT, the same bit
that is used by the ctrl-C handler. Use a separate bit so that
this sw-generated abort does not cause less to exit if -K is set.

Related to #628.