build: Make mingw build work properly from autotools
This change makes a mingw build work from the standard autotools ./configure
workflow. There are two mostly independent changes:
1. Change the mingw detection preprocessor macro to the standard __MINGW32__
   which is defined by the compiler (rather than expecting MINGW to be set by
   the defines header).
2. In the autotools-emitted defines.h make the values of MSDOS_COMPILER and
   PATHNAME_SEP appropriately automatically.

I understand that there already is a spearate Makefile for mingw. However,
it would be convenient to make the autotools version just work as well.
In paprticular, I'm working on packaging less for https://github.com/JuliaPackaging/Yggdrasil.
The system sets various defaults when it detects an autotools build.
Of course, it would be possible to manually apply these for a mingw
build, but it would be easier if the autotools build just worked.