Fix hang with ^S modifier if search matches empty string.
b71a5788d added a loop in match_pattern that iterates through
matches in a line looking for one that also satisfies any
SUBSEARCH requirement in effect (via the ^S modifier).
It starts each search at the end of the string matched by
the previous search. But if a search matches an empty string,
this will get stuck in the loop, continually searching from the same point.

Related to #634.