Fix bug when pasting a search pattern while --incsearch is enabled.
After receiving each character of the pattern we call search() to
find matches. When search calls down to iread(), the pending pasted
input can cause iread to abort and return READ_INTR which ends the
search. This seems to happen only when input is a pipe.

Fix by disabling polling during this call to search().

This has a risk of getting into a state where iread() is blocked
waiting for data but is not interruptible by ^X or other non-^C
chars, but I have not been able to produce such a case.

Related to #635.