Improve performance with very long lines.
Previously get_forw_line() was by design stateless. So if a very
long file line were displayed, each call to get_forw_line() to
get a screen line would scan backwards to the beginning of the file
line, then scan forward again back to the screen line position.
This was mainly to process any SGR sequences and make sure we have
the correct text color at the start of the screen line.

This change remembers the position of each screen line processed
by get_forw_line(). If the next call to get_forw_line() asks for the
next contiguous screen line, we obtain the SGR sequences from the
previous line buffer so we don't need to scan backwards and forwards.

Related to #621.