Improve behavior of --incsearch.
Previously with --incsearch enabled, after each character of a search
pattern was entered, less would search for the pattern starting at the
current position. This could cause some matches to be missed if there
was a match for a shorter substring of the pattern later in the file.
This change makes each of the per-char searches start at the same
position, the position at which the search command was first invoked.

Related to #640.