Fix problem assigning the "invalid" in a lesskey file.
The "invalid" action translates to A_UINVALID, which was being
treated as if it were A_INVALID. This causes the "invalid" action
in a user-defined lesskey table to be ignored and the predefined
action, if any, was used. Instead, A_UINVALID should be treated
as a unique command distinct from A_INVALID.

The treating of A_UINVALID as A_INVALID dates back to
fb62f1f51860e18dfa755f2739af8e781aa4facc from 9 July 2000.
The commit message says it is to fix a bug using #stop, but I
don't see any bugs now and I don't understand why it was added.
The command decoder has been substantially rewritten since then,
so I assume this is not needed any more, although I would feel
better about this change if I understood the original motivation.

Related to #643.