Don't set EOF upon receiving an abort signal.
Previously if an abort signal was received while reading the input
file, we would set the EOF position to the current read position.
I'm not sure why this was being done. I can't see any advantage,
and it causes the file to end prematurely when the user sends a
^X or ^C, or (since 23ff6a42488a9eee83341265c0c053d97b56f33d), if
the user hits any key.

Also suppress the eof_bell if forw() is interrupted by a signal.
This is more covering up an issue than fixing it, since we could
still exit forw() early when receiving an input char during the
file read, but at least it doesn't ring the bell to call attention
to the early abort now. I'm not sure how to truly avoid the early
abort while still allowing any key to interrupt a file read; these
seem contradictory goals.

Related to #649.