Change behavior of HOME & END, and implement shift and ctrl versions.
Previous HOME was like g (go to beginning of file) and
END was like G (go to end of file).
Change HOME to be like ESC-{ (go to beginning of line) and
change END to be like ESC-} (go to end of line).
This is more consistent with other applications.

Also implement ctl-HOME and shift-HOME to be like g-ESC-{
(go to beginning of file and beginning of line) and
implement ctl-END and shift-END to be like G-ESC-}
(go to end of file and end of line).

The ctl- versions depend on the terminfo entry for the terminal
containing the nonstandard kHOM5 and kEND5 caps.

Related to #658.