Prevent URLs from oversetting line if possible.
AT&T troff had no means of designating break points in words that should
not be hyphenated; the program was developed in an era when file names
were 14 characters long at most and URLs were over a decade away.

groff, Heirloom Doctools troff, and mandoc support the `\:` escape
sequence, which solves the problem.

Adapt a technique from ncurses[1], using a string to access this escape
sequence only if the formatter claims groff compatibility.

URLs continue to be at risk of oversetting on other formatters.

Also sync the bug-reporting advice in lessecho(1) and lesskey(1) with
that of less(1).

Further, I would not count on any formatter to respect the technique
you're using to prevent URL hyphenation by bracketing the URL with `nh`
and `hy` requests.  When the formatter reaches the end of an output line
and the hyphenation mode is nonzero _at that time_, it marches backwards
through the line and hyphenates accordingly.  This is a known defect of
both GNU and AT&T troffs.[2]  Worse, because the `hy` request without
arguments sets the hyphenation mode to "1", this approach defeats the
preferences of any groff users specifying `-rHY=0`, a feature supported
since groff 1.19 (April 2003).

[1] https://github.com/ThomasDickey/ncurses-snapshots/blob/4311b18dda3e30eed79bb7dca58cb52bab70c942/man/curs_getch.3x#L46
    https://github.com/ThomasDickey/ncurses-snapshots/blob/4311b18dda3e30eed79bb7dca58cb52bab70c942/man/curs_getch.3x#L782
[2] https://savannah.gnu.org/bugs/?67310
[3] https://www.gnu.org/software/groff/manual/groff.html.node/Manipulating-Hyphenation.html

    When trying to write documents portable to AT&T troff, the problems
    are worse.

    https://cgit.git.savannah.gnu.org/cgit/groff.git/tree/tmac/an.tmac?id=15ab324cdbf9a727094a8385e6d902077405ff6e#n199

Here's a diff of the impact of this change, using
`nroff -dAD=l -ww -t -man -P -cbou *.nro`.

$ diff -U1 /tmp/less[12]
--- /tmp/less1  2025-10-11 10:44:30.701006521 -0500
+++ /tmp/less2  2025-10-11 10:44:42.600979073 -0500
@@ -762,4 +762,4 @@
                    but instead as a 4‐bit CHAR_INFO.Attributes value, between 0
-                   and 15 inclusive (see
-                   https://learn.microsoft.com/en-us/windows/console/char-info-str).
+                   and 15 inclusive (see https://learn.microsoft.com/en-us/
+                   windows/console/char-info-str).

@@ -1072,7 +1072,7 @@
             is taken to be the name of a command compatible with global(1), and
-            that command is executed to find the tag.  (See
-            http://www.gnu.org/software/global/global.html).  The -t option may
-            also be specified from within less (using the - command) as a way of
-            examining a new file.  The command ":t" is equivalent to specifying
-            -t from within less.
+            that command is executed to find the tag.  (See http://www.gnu.org/
+            software/global/global.html).  The -t option may also be specified
+            from within less (using the - command) as a way of examining a new
+            file.  The command ":t" is equivalent to specifying -t from within
+            less.

@@ -1438,4 +1438,5 @@
             to change the colors.  Colored text works only if the terminal sup‐
-            ports ANSI color escape sequences (as defined in
-            https://www.ecma-international.org/publications-and-standards/standards/ecma-48).
+            ports ANSI color escape sequences (as defined in https://www
+            .ecma-international.org/publications-and-standards/standards/
+            ecma-48).