Make ESC-u clear OSC 8 highlight without error message.
Previously if there was an OSC 8 link highlighted but no search
pattern in effect, the ESC-u command would internally clear the
OSC 8 highlight, but not erase it from the screen, and would also
display a "No previous regular expression" message.
This change fixes both problems; the display is updated and
no error message is displayed.