Fix bug in -j option on command line.
We need to keep the argument of the -j option separate from the actual
value used internally. After we recalculated jump_sline based on the
screen size, we must retain the value of the -j parameter so that it
can be recalculated correctly when calc_jump_sline is called again.