Don't initialize terminal until input data has been read.
Defer sending init sequences to the terminal until the first byte
of input data has been read. On alt-screen terminals, this avoids
switching to the alternate screen immediately, which in some scenarios
can be useful. For example, it allows any initial stderr messages from
a piping program to be seen.

Related to #682.