Allow line numbers to be colored without bold formatting.
Previously the attribute for line numbers was set to
(AT_BOLD|AT_COLOR_LINENUM). So changing the color of
AT_COLOR_LINENUM couldn't remove the AT_BOLD attribute.

Instead, set the attribute to
(use_color ? AT_COLOR_LINENUM : AT_BOLD), and the color string
for AT_COLOR_LINENUM to "c*". This makes the formatting bold
if --color is not set, and cyan+bold if --color is set, but
changing the color string for -DN can remove the bold.

Related to #684.