Don't perform terminal operations before calling term_init.
In a55140f63f409cf0069279a6dbaf2d0ea385a2e5 we defer term_init until
receiving the first byte of file data. But in some cases we perform
terminal operations like clearing the screen before that point.
This is bad if, for example, term_init switches to an alt screen;
then we are clearing the wrong screen. The two cases fixed in this
commit are:
1. Using -c, which clears the screen before the first screen display.
2. Setting LESS_LINES which causes make_display to clear the screen
   before data is received.
After a code review these are the only examples I have found.

Related to #686.