Fix crash in "less --version".
After setting output to stdout, set term_init_ever to prevent
putchr() from initializing the terminal. This needs to be done
when using stdout for output other than the usual file contents
(currently when catting to a non-tty, and in printing --version output).
Bundle this into set_output to make the requirement more clear.

Related to #698.