Don't disable features based on absence of strstr().
defines.wn was erroneously missing a definition for HAVE_STRSTR, making
less think strstr was unavailable. This made some features unavailable
on the Windows build, notably support for vim-compatible modelines.

This commit makes two changes: it adds a definition for HAVE_STRSTR
in defines.wn (as well as a couple of other missing definitions), and
it provides a local implementation of strstr() for systems where strstr
really is not implemented by the system.