Reorganize defines.wn.
Reorder entries to be closer to defines.h generated by autoconf.
Also add some #defines to explicitly set symbols to 0 instead of
relying on their absence.
This makes it easier to compare defines.h to defines.wn and see
if anything is missing.

This should probably be done for the other hardcoded defines files
too (defines.ds, defines.o2, defines.o9) but it's too dangerous
to change them without access to a compiler and runtime.