Fix screen resize on Windows.
Previously on Windows, after resizing the console window, the content
would not be repainted to use the new window size until the next command
key was received. Also, not everything dependent on the screen size was
correctly recalculated after the size change. For example, if you
set --shift=.5 to make the horizontal shift width be half of the screen
width, the shift width would not be recalculated after a window resize.

This fixes the recalculation issues, and also makes the screen repaint
immediately after the window is resized.