Some fixes for recent regressions in Windows console resizing.
It seems to be necessary to call term_init() at startup rather than
deferring to the first putchr as in 67202cb530958a798bd44c39ca4310710a2ee769.
Without this, any char in the rightmost column is not displayed. I think this
is related to the setting of ENABLE_WRAP_AT_EOL_OUTPUT in win32_init_vt_term
but I don't fully understand it.

It is also necessary to call check_winch() in commands(). Without this, the
rightmost char is also not displayed. So in win32_window_event, don't set
sigs to call psignals, just enqueue a READ_AGAIN which will cause check_winch
to be called when we return to commands(). It does not work to call
check_winch in win32_window_event or in psignals; it has to be done in
commands(). I understand this even less than the above term_init issue but it
seems to be related to some behavior of the Windows console that I don't want
to spend any more time investigating. Since check_winch has been called from
commands() for years, I think it's safest to just restore it and rely on it.